refactor: 포맷팅, 출력 문구 수정

+++ src/main/java/lotto/domain/Lottos.java
@@ -70,7 +70,7 @@ public class Lottos {
             totalPrize += (prize * count);
         }
 
-        float size = lottos.size();
+        int size = lottos.size();
         float totalBuy = size * 1000;
         this.totalRevenue = totalPrize / totalBuy;
         return totalRevenue;
